+++ core/src/sn/ScriptIterator.ts
@@ -9,7 +9,7 @@ import {uint, argChk_Boolean, getFn, CmnLib} from './CmnLib';
 import {IHTag, IMain, IVariable, IMark, HArg, Script, IPropParser} from './CmnInterface';
 import {Config} from './Config';
 import {CallStack, ICallStackArg} from './CallStack';
-import {Grammar, REG_TAG, REG_TAG_NAME} from './Grammar';
+import {Grammar, tagToken2Name_Args, tagToken2Name} from './Grammar';
 import {AnalyzeTagArg} from './AnalyzeTagArg';
 
 import {EventMng} from './EventMng';
@@ -273,9 +273,7 @@ export class ScriptIterator {
 
 			default:
 			{	// 関数ブレークポイント
-				const e = REG_TAG_NAME.exec(token);
-				const tag_name = e?.groups?.name ?? '';
-				if (tag_name in ScriptIterator.hFuncBP) {
+				if (tagToken2Name(token) in ScriptIterator.hFuncBP) {
 					this.breakState = BreakState.break;
 					this.main.setLoop(false, `関数 ${token} ブレーク`);
 					this.sys.callHook('stopOnBreakpoint', {});	// sn全体へ通知
@@ -312,10 +310,8 @@ export class ScriptIterator {
 		const idx_n = this.breakState === BreakState.breaking ?1 :0;
 		const tkn0 = this.script.aToken[this.idxToken_ -1 +idx_n];
 
-		const e0 = REG_TAG_NAME.exec(tkn0);
-		const tag_name0 = e0?.groups?.name ?? '';
-
 		const fn0 = this.cnvSnPath4Dbg(this.scriptFn_);
+		const tag_name0 = tagToken2Name(tkn0);
 		const nm = tag_name0 ?`[${tag_name0}]` :tkn0;
 //console.log(`fn:ScriptIterator.ts line:425 aStack breakState:${this.breakState} idx:${this.idxToken_ -1} idx_n:${idx_n} tkn0:${tkn0}: nm:${nm} tkn02:${this.script.aToken[this.idxToken_ -1]}: +tkn02:${this.script.aToken[this.idxToken_]}:`);
 //console.log(`fn:ScriptIterator.ts line:426    a:%o anum:%o`, this.script.aToken, this.script.aLNum);
@@ -334,8 +330,8 @@ export class ScriptIterator {
 			const st = this.hScript[cs.fn];
 			const tkn = st.aToken[cs.idx -1];
 			const lc = this.cnvIdx2lineCol(st, cs.idx);	// -1不要
-			const e = REG_TAG_NAME.exec(tkn);
-			const tag_name = e?.groups?.name ?? '';
+
+			const tag_name = tagToken2Name(tkn);
 			a.push({
 				fn: this.cnvSnPath4Dbg(cs.fn),
 				ln: lc.ln,
@@ -350,15 +346,11 @@ export class ScriptIterator {
 	// result = true : waitする  resume()で再開
 	private	procDebugtag	= (_tag_name: string)=> {};
 	タグ解析(tagToken: string): boolean {
-		const e = REG_TAG.exec(tagToken);
-		const g = e?.groups;
-		if (! g) throw `タグ記述【${tagToken}】異常です(タグ解析)`;
-
-		const tag_name = g.name;
+		const [tag_name, args] = tagToken2Name_Args(tagToken);
 		const tag_fnc = this.hTag[tag_name];
 		if (! tag_fnc) throw `未定義のタグ【${tag_name}】です`;
 
-		this.alzTagArg.go(g.args);
+		this.alzTagArg.go(args);
 		this.procDebugtag(tag_name);
 
 		const hPrm = this.alzTagArg.hPrm;
@@ -597,12 +589,9 @@ export class ScriptIterator {
 			if (uc === 10) {this.addLineNum(t.length); continue;}	// \n 改行
 			if (uc !== 91) continue;		// [ タグ開始以外
 
-			const a_tag = REG_TAG.exec(t);
-			const g = a_tag?.groups;
-			if (! g) throw 'タグ記述['+ t +']異常です(if文)';
-			const tag_name = g.name;
+			const [tag_name, args] = tagToken2Name_Args(t);
 			if (! (tag_name in this.hTag)) throw `未定義のタグ[${tag_name}]です`;
-			this.alzTagArg.go(g.args);
+			this.alzTagArg.go(args);
 
 			switch (tag_name) {
 			case 'if':	++cntDepth; break;
@@ -949,17 +938,15 @@ export class ScriptIterator {
 			this.REG_WILDCARD.lastIndex = 0;
 			if (! this.REG_WILDCARD.test(token)) continue;
 
-			const e = REG_TAG.exec(token);
-			const g = e?.groups;
-			if (! g) continue;
-			this.alzTagArg.go(g.args);
+			const [tag_name, args] = tagToken2Name_Args(token);
+			this.alzTagArg.go(args);
 
 			const p_fn = this.alzTagArg.hPrm.fn;
 			if (! p_fn) continue;
 			const fn = p_fn.val;
 			if (! fn || fn.slice(-1) !== '*') continue;
 
-			const ext = (g.name === 'loadplugin') ?'css' :'sn';
+			const ext = (tag_name === 'loadplugin') ?'css' :'sn';
 			const a = this.cfg.matchPath('^'+ fn.slice(0, -1) +'.*', ext);
 
 			this.script.aToken.splice(i, 1, '\t', '; '+ token);